Move out navigation links, use NavigationHeader in signin/login page
