+++ src/routes/__layout-root.svelte
@@ -73,10 +73,10 @@
 <main class="flex flex-col h-screen">
   <Notifications />
   <Banner {cluster} />
-  <Header {user} />
-  <HeaderResponsive {user} />
+  <Header user={me} />
+  <HeaderResponsive user={me} />
   <section id="content" class="h-full mx-10 mb-10 mt-8">
-    <div class="flex flex-col h-full gap-4">
+    <div class="flex flex-col h-full gap-4 mt-12 xl:mt-0">
       <ErrorBoundary onError={() => {}}>
         <slot />
       </ErrorBoundary>
