+++ src/routes/import/__layout-import.svelte
@@ -49,7 +49,7 @@
   <Header {user} />
   <HeaderResponsive {user} />
   <section id="content" class="h-full mx-10 mb-10 mt-8">
-    <div class="flex flex-col h-full gap-4">
+    <div class="flex flex-col h-full gap-4 mt-12 xl:mt-0">
       <ErrorBoundary onError={() => {}}>
         <slot />
       </ErrorBoundary>
